### 2022-11-25 #4966 @arfon
@5uperpalo – looks like one of your authors has an invalid ORCID here. Could you please fix and try compiling the paper again with the command `@editorialbot generate pdf`: https://github.com/openjournals/joss-papers/actions/runs/3546332228/jobs/5955266136#step:3:97
