### 2022-11-27 #4966 @arfon
Output from `cloc`:

```
cloc pytorch-widedeep/
     370 text files.
     342 unique files.
      50 files ignored.

Language                     files          blank        comment           code
HTML                            38          36312            143          72666
Python                         155           4051           6185          18354
Jupyter Notebook                48              1          36483           6116
JavaScript                      31           1021           1324           4963
Markdown                        23            411              0           1360
YAML                             4             19             35            266
reStructuredText                18            186            307            242
XML                              1              0              0            183
CSS                              6             30             21            155
DOS Batch                        1              8              1             26
make                             1              4              7             10
Bourne Shell                     1              0              4              4
JSON                             1              0              0              1
SUM:                           328          42043          44510         104346
```
