### 2022-12-06 00759d4
Added script with multiclass using text and tabular. Simple tf idf and lightgbm
