+++ examples/scripts/ecomerce_reviews.py
@@ -1,0 +1,169 @@
+import numpy as np
+import lightgbm as lgb
+from lightgbm import Dataset as lgbDataset
+from scipy.sparse import hstack, csr_matrix
+from sklearn.metrics import (
+    f1_score,
+    recall_score,
+    accuracy_score,
+    precision_score,
+    confusion_matrix,
+)
+from sklearn.model_selection import train_test_split
+from sklearn.feature_extraction.text import TfidfVectorizer
+
+from pytorch_widedeep.utils import Tokenizer, LabelEncoder
+from pytorch_widedeep.datasets import load_womens_ecommerce
+
+df = load_womens_ecommerce(as_frame=True)
+
+df.columns = [c.replace(" ", "_").lower() for c in df.columns]
+
+# classes from [0,num_class)
+df["rating"] = (df["rating"] - 1).astype("int64")
+
+# group reviews with 1 and 2 scores into one class
+df.loc[df.rating == 0, "rating"] = 1
+
+# and back again to [0,num_class)
+df["rating"] = (df["rating"] - 1).astype("int64")
+
+# drop short reviews
+df = df[~df.review_text.isna()]
+df["review_length"] = df.review_text.apply(lambda x: len(x.split(" ")))
+df = df[df.review_length >= 5]
+df = df.drop("review_length", axis=1).reset_index(drop=True)
+
+train, valid = train_test_split(df, train_size=0.8, random_state=1, stratify=df.rating)
+valid, test = train_test_split(
+    valid, train_size=0.5, random_state=1, stratify=valid.rating
+)
+
+
+###############################################################################
+# The simplest thing: tokenization + tf-idf
+tok = Tokenizer()
+tok_reviews_tr = tok.process_all(train.review_text.tolist())
+tok_reviews_va = tok.process_all(valid.review_text.tolist())
+vectorizer = TfidfVectorizer(
+    max_features=5000, preprocessor=lambda x: x, tokenizer=lambda x: x, min_df=5
+)
+
+X_text_tr = vectorizer.fit_transform(tok_reviews_tr)
+X_text_va = vectorizer.transform(tok_reviews_va)
+
+lgbtrain_text = lgbDataset(
+    X_text_tr,
+    train.rating.values,
+    free_raw_data=False,
+)
+
+lgbvalid_text = lgbDataset(
+    X_text_va,
+    valid.rating.values,
+    reference=lgbtrain_text,
+    free_raw_data=False,
+)
+
+model_text = lgb.train(
+    {"objective": "multiclass", "num_classes": 4},
+    lgbtrain_text,
+    valid_sets=[lgbvalid_text, lgbtrain_text],
+    valid_names=["test", "train"],
+    early_stopping_rounds=50,
+)
+
+preds_text = model_text.predict(X_text_va)
+pred_text_class = np.argmax(preds_text, 1)
+
+acc_text = accuracy_score(lgbvalid_text.label, pred_text_class)
+f1_text = f1_score(lgbvalid_text.label, pred_text_class, average="weighted")
+prec_text = precision_score(lgbvalid_text.label, pred_text_class, average="weighted")
+rec_text = recall_score(lgbvalid_text.label, pred_text_class, average="weighted")
+cm_text = confusion_matrix(lgbvalid_text.label, pred_text_class)
+
+
+###############################################################################
+# Text + Tabular
+
+tab_cols = [
+    "age",
+    "division_name",
+    "department_name",
+    "class_name",
+]
+
+for tab_df in [train, valid]:
+    for c in ["division_name", "department_name", "class_name"]:
[…]
