### 2022-12-07 f8ff348
close to finishing a HG example
