### 2022-12-14 #4966 @arfon
@5uperpalo - looks like that ORCID is working now, that you. BTW, we're currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.
