### 2022-12-16 2593954
Added example on how to use a HF model
