### 2022-12-17 fb4c930
fixed a bug related to using a FC-head in top of a model via the WideDeep class
