### 2022-12-20 #4966 @osorensen
👋 @aadi-mishra @dataplayer12 @siboehm, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
