### 2022-12-23 #5027 @editorialbot
[REVIEW]: pytorch-widedeep: A flexible package for multimodal-deep-learning
**Submitting author:** @5uperpalo (<a href="http://orcid.org/0000-0002-9394-8794">Pavol Mulinka</a>)
**Repository:** https://github.com/jrzaurin/pytorch-widedeep
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** 1.2.0
