### 2022-12-24 5fecd7a
notebook on HF models almost finished. It wont run in bloody jupyter notebooks but runs perfectly as script
