### 2022-12-26 #5027 @makoeppel
Following line 50 of your paper, can you briefly explain what \sigma and b are?
