### 2022-12-26 #5027 @5uperpalo
@makoeppel first of all, thank you for doing the review... from the cited paper(see attached screenshot, @jrzaurin correct me if in case I am wrong):
```
σ(·) is the sigmoid function
b is the bias term
```
**question** - should we add the explanation to the JOSS paper? or wait for possible other comments and add it afterwards?
![image](https://user-images.githubusercontent.com/36483359/209561487-bfad8545-8ada-4b83-bc4f-ffb968c0624d.png)
