### 2022-12-26 jrzaurin/pytorch-widedeep#117 @makoeppel
Page not found error for CONTRIBUTING
Imam getting a page not found error when clicking on "Check CONTRIBUTING page." In the readme.
