### 2022-12-26 #5027 @makoeppel
Contributing link in the readme gets an page not found error: https://github.com/jrzaurin/pytorch-widedeep/issues/117
