### 2022-12-26 #5027 @makoeppel
Can you maybe compare your framework to the Wide & Deep / Deep & Cross models from Keras / Tensorflow? Here are some (incomplete) links to them:

- https://www.tensorflow.org/api_docs/python/tf/keras/experimental/WideDeepModel
- https://keras.io/examples/structured_data/wide_deep_cross_networks/
- https://ai.googleblog.com/2016/06/wide-deep-learning-better-together-with.html?m=1
