### 2022-12-26 #5027 @osorensen
> @makoeppel first of all, thank you for doing the review... from the cited paper(see attached screenshot, @jrzaurin please correct me if I am wrong):
>
> ```
> σ(·) is the sigmoid function
> b is the bias term
> ```
>
> **question** - should we add the explanation to the JOSS paper? or wait for possible other comments and add it afterwards? ![image](https://user-images.githubusercontent.com/36483359/209561487-bfad8545-8ada-4b83-bc4f-ffb968c0624d.png)

Nice to see the review has already started!

Regarding your **question** @5uperpalo, as editor I'd just like the chime in that given the interactive nature of the review process, I think editing this already now sounds like a good idea.
