### 2022-12-29 2bc2742
tiny changes related to JOSS review
