+++ mkdocs/paper_JOSS/paper.bib
@@ -173,6 +173,7 @@ doi = {10.18653/v1/N16-1174}
 }
 
 @inproceedings{he2016deep,
+  doi= {10.1109/cvpr.2016.90},
   title={Deep residual learning for image recognition},
   author={He, Kaiming and Zhang, Xiangyu and Ren, Shaoqing and Sun, Jian},
   booktitle={Proceedings of the IEEE conference on computer vision and pattern recognition},
@@ -182,6 +183,7 @@ doi = {10.18653/v1/N16-1174}
 
 
 @inproceedings{zhang2018shufflenet,
+  doi={10.1109/CVPR.2018.00716},
   title={Shufflenet: An extremely efficient convolutional neural network for mobile devices},
   author={Zhang, Xiangyu and Zhou, Xinyu and Lin, Mengxiao and Sun, Jian},
   booktitle={Proceedings of the IEEE conference on computer vision and pattern recognition},
@@ -190,6 +192,7 @@ doi = {10.18653/v1/N16-1174}
 }
 
 @inproceedings{xie2017aggregated,
+  doi={10.1109/CVPR.2017.634},
   title={Aggregated residual transformations for deep neural networks},
   author={Xie, Saining and Girshick, Ross and Doll{\'a}r, Piotr and Tu, Zhuowen and He, Kaiming},
   booktitle={Proceedings of the IEEE conference on computer vision and pattern recognition},
@@ -198,6 +201,7 @@ doi = {10.18653/v1/N16-1174}
 }
 
 @article{zagoruyko2016wide,
+  doi={10.5244/c.30.87},
   title={Wide residual networks},
   author={Zagoruyko, Sergey and Komodakis, Nikos},
   journal={arXiv preprint arXiv:1605.07146},
@@ -205,6 +209,7 @@ doi = {10.18653/v1/N16-1174}
 }
 
 @article{xu2022regnet,
+  doi={10.1109/TNNLS.2022.3158966},
   title={RegNet: self-regulated network for image classification},
   author={Xu, Jing and Pan, Yu and Pan, Xinglin and Hoi, Steven and Yi, Zhang and Xu, Zenglin},
   journal={IEEE Transactions on Neural Networks and Learning Systems},
@@ -228,6 +233,7 @@ doi = {10.18653/v1/N16-1174}
 }
 
 @inproceedings{tan2019mnasnet,
+  doi={10.1109/CVPR.2019.00293},
   title={Mnasnet: Platform-aware neural architecture search for mobile},
   author={Tan, Mingxing and Chen, Bo and Pang, Ruoming and Vasudevan, Vijay and Sandler, Mark and Howard, Andrew and Le, Quoc V},
   booktitle={Proceedings of the IEEE/CVF Conference on Computer Vision and Pattern Recognition},
@@ -263,6 +269,7 @@ doi = {10.1162/neco.1997.9.8.1735}
 }
 
 @article{cho2014properties,
+  doi={10.3115/v1/w14-4012},
   title={On the properties of neural machine translation: Encoder-decoder approaches},
   author={Cho, Kyunghyun and Van Merri{\"e}nboer, Bart and Bahdanau, Dzmitry and Bengio, Yoshua},
   journal={arXiv preprint arXiv:1409.1259},
