### 2023-01-05 104a763
completed the example notebook and fixed a few issues related to newer pytorch versions
