+++ mkdocs/paper_JOSS/paper.bib
@@ -97,6 +97,7 @@ url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-perf
 
 
 @inproceedings{cheng2016wide,
+  doi={10.48550/arXiv.1606.07792},
   title={Wide \& deep learning for recommender systems},
   author={Cheng, Heng-Tze and Koc, Levent and Harmsen, Jeremiah and Shaked, Tal and Chandra, Tushar and Aradhye, Hrishi and Anderson, Glen and Corrado, Greg and Chai, Wei and Ispir, Mustafa and others},
   booktitle={Proceedings of the 1st workshop on deep learning for recommender systems},
@@ -366,4 +367,13 @@ doi = {10.1162/neco.1997.9.8.1735}
   booktitle={Proceedings of the Thirteenth Language Resources and Evaluation Conference},
   pages={6837--6847},
   year={2022}
+}
+
+@incollection{wang2017deep,
+ 	doi={10.48550/arXiv.1708.05123},
+  title={Deep \& cross network for ad click predictions},
+  author={Wang, Ruoxi and Fu, Bin and Fu, Gang and Wang, Mingliang},
+  booktitle={Proceedings of the ADKDD'17},
+  pages={1--7},
+  year={2017}
 }
