### 2023-01-08 c3e0563
fixed torchmetrics test
