### 2023-01-08 2dcd688
fixed torchmetrics test
