+++ tests/test_metrics/test_torchmetrics.py
@@ -34,7 +34,7 @@ y_pred_bin_pt = torch.from_numpy(y_pred_bin_np)
         ("Precision", precision_score, Precision(task="binary")),
         ("Recall", recall_score, Recall(task="binary")),
         ("F1Score", f1_score, F1Score(task="binary")),
-        ("FBetaScore", f2_score_bin, FBetaScore(beta=2)),
+        ("FBetaScore", f2_score_bin, FBetaScore(task="binary", beta=2)),
     ],
 )
 def test_binary_metrics(metric_name, sklearn_metric, torch_metric):
@@ -77,10 +77,26 @@ def f2_score_multi(y_true, y_pred, average):
 @pytest.mark.parametrize(
     "metric_name, sklearn_metric, torch_metric",
     [
-        ("Accuracy", accuracy_score, Accuracy(task="multiclass", num_classes=3, average="micro")),
-        ("Precision", precision_score, Precision(task="multiclass", num_classes=3, average="macro")),
-        ("Recall", recall_score, Recall(task="multiclass", num_classes=3, average="macro")),
-        ("F1Score", f1_score, F1Score(task="multiclass", num_classes=3, average="macro")),
+        (
+            "Accuracy",
+            accuracy_score,
+            Accuracy(task="multiclass", num_classes=3, average="micro"),
+        ),
+        (
+            "Precision",
+            precision_score,
+            Precision(task="multiclass", num_classes=3, average="macro"),
+        ),
+        (
+            "Recall",
+            recall_score,
+            Recall(task="multiclass", num_classes=3, average="macro"),
+        ),
+        (
+            "F1Score",
+            f1_score,
+            F1Score(task="multiclass", num_classes=3, average="macro"),
+        ),
         (
             "Accuracy",
             accuracy_score,
@@ -124,7 +140,7 @@ def f2_score_multi(y_true, y_pred, average):
         (
             "MulticlassFBetaScore",
             f2_score_multi,
-            FBetaScore(task="multiclass", beta=3.0, num_classes=3, average="macro"),
+            FBetaScore(task="multiclass", beta=3, num_classes=3, average="macro"),
         ),
     ],
 )
