### 2023-01-11 jrzaurin/pytorch-widedeep#117 @5uperpalo
Hi @makoeppel , it too me a bit more time as @jrzaurin also added an example with hugging face integration to the library etc... Can I close this issue ? it is fixed in contribution on github and in readthedocs:
* https://pytorch-widedeep.readthedocs.io/en/latest/contributing.html
* https://github.com/jrzaurin/pytorch-widedeep/blob/master/CONTRIBUTING.MD
