### 2023-01-11 jrzaurin/pytorch-widedeep#124 @5uperpalo
[JOSS review] Can you maybe compare your framework to the Wide & Deep / Deep & Cross models from Keras / Tensorflow?
related to https://github.com/openjournals/joss-reviews/issues/5027#issuecomment-1365260401
