### 2023-01-11 jrzaurin/pytorch-widedeep#124 @5uperpalo
Hi @makoeppel can I close this issue related to JOSS paper review? I added a note comparing our library to keras/tensorflow implementation, here:
* https://github.com/jrzaurin/pytorch-widedeep/blob/joss_paper/mkdocs/paper_JOSS/paper.md#statement-of-need
```
As opposed to Google's "Wide and Deep" and "Deep and Cross"[@wang2017deep] architecture implementations in Keras/Tensorflow, we use the wide/cross and deep model design as an initial building blocks of pytorch deep learning models to provide basis for plethora of state-of-the-art models and architecture implementations that can be seamlessly assembled with just a few lines of code. Additionally, the individual components do not necessarly have to be a part of the final architecture.
```
