### 2023-01-11 jrzaurin/pytorch-widedeep#117 @makoeppel
Hi @5uperpalo, thank you for fixing this. I have one problem with the Roadmap link on the contributing.html/md. This link is not working for me.
