### 2023-01-11 #5027 @makoeppel
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.48550/ARXIV.2003.06505 is OK
> - 10.48550/ARXIV.2108.09084 is OK
> - 10.18653/v1/N16-1174 is OK
> - 10.1162/neco.1997.9.8.1735 is OK
>
> MISSING DOIs
>
> - 10.1109/cvpr.2016.90 may be a valid DOI for title: Deep residual learning for image recognition
> - 10.1109/cvpr.2018.00716 may be a valid DOI for title: Shufflenet: An extremely efficient convolutional neural network for mobile devices
> - 10.1109/cvpr.2017.634 may be a valid DOI for title: Aggregated residual transformations for deep neural networks
> - 10.5244/c.30.87 may be a valid DOI for title: Wide residual networks
> - 10.1109/tnnls.2022.3158966 may be a valid DOI for title: RegNet: self-regulated network for image classification
> - 10.1109/cvpr.2019.00293 may be a valid DOI for title: Mnasnet: Platform-aware neural architecture search for mobile
> - 10.3115/v1/w14-4012 may be a valid DOI for title: On the properties of neural machine translation: Encoder-decoder approaches
>
> INVALID DOIs
>
> - None
> ```

Hi @5uperpalo, as you can see the bot shows some missing DOIs for the references. Can you check this? Otherwise I am happy with the current status of the paper.md.
