### 2023-01-11 #5027 @makoeppel
> @makoeppel , this is strange, I added al of those DOIs after I noticed this message (1-2 weeks ago?), they are all in paper.bib, any idea why is this still showing up?
>
>     * https://github.com/jrzaurin/pytorch-widedeep/blob/d05352c73283f6b830225976f6bb7b7699fc658e/mkdocs/paper_JOSS/paper.bib#L177
>
>     * https://github.com/jrzaurin/pytorch-widedeep/blob/d05352c73283f6b830225976f6bb7b7699fc658e/mkdocs/paper_JOSS/paper.bib#L187
>
>     * https://github.com/jrzaurin/pytorch-widedeep/blob/d05352c73283f6b830225976f6bb7b7699fc658e/mkdocs/paper_JOSS/paper.bib#L196
>
>     * https://github.com/jrzaurin/pytorch-widedeep/blob/d05352c73283f6b830225976f6bb7b7699fc658e/mkdocs/paper_JOSS/paper.bib#L205
>
>     * https://github.com/jrzaurin/pytorch-widedeep/blob/d05352c73283f6b830225976f6bb7b7699fc658e/mkdocs/paper_JOSS/paper.bib#L213
>
>     * https://github.com/jrzaurin/pytorch-widedeep/blob/d05352c73283f6b830225976f6bb7b7699fc658e/mkdocs/paper_JOSS/paper.bib#L237
>
>     * https://github.com/jrzaurin/pytorch-widedeep/blob/d05352c73283f6b830225976f6bb7b7699fc658e/mkdocs/paper_JOSS/paper.bib#L273

@5uperpalo sorry this was on my end I did not notice that you changed them already. However, there is still one missing.
