+++ mkdocs/paper_JOSS/paper.bib
@@ -342,6 +342,7 @@ doi = {10.1162/neco.1997.9.8.1735}
 }
 
 @manual{keras,
+  doi={10.1163/1574-9347_dnp_e612900},
   title={keras},
   note="[Online]. Available at: https://keras.io/"
 }
