### 2023-01-22 jrzaurin/pytorch-widedeep#127 @siboehm
Licences of copied code
As stated in the [JOSS Paper](https://github.com/openjournals/joss-reviews/issues/5027), parts of the pytorch-widedeep packages are copied from other repositories. One example: [fastai_transforms](https://github.com/jrzaurin/pytorch-widedeep/blob/master/pytorch_widedeep/utils/fastai_transforms.py) contains code copied from [fastai/text/core.py](https://github.com/fastai/fastai/blob/dedec6be9334b1c3abfcdc9a70fe43898a74e39e/fastai/text/core.py). But fast.ai is licensed under Apache-2.0, while this package is licensed under MIT. Do these two go together?

At the very least the copied code should be properly credited, including the original license (same for the other copied parts).
