### 2023-01-22 jrzaurin/pytorch-widedeep#127 @jrzaurin
I am actually not 100% sure, we'll check. The fastai code is credited in our code, and if it is not in the paper is our mistake, will fix it asap.
