### 2023-01-22 jrzaurin/pytorch-widedeep#127 @siboehm
Sounds good! Fastai is credited in the paper, that's how I found out about it :) I'm just worried about A) how apache and MIT go together and b) whether or not you need to include the original apache license text in `fastai_transforms.py` (instead of just the credit comment).
