### 2023-01-23 #5027 @osorensen
Thanks @siboehm! @5uperpalo, please let us know in this thread when you've fixed the issues pointed out by @siboehm. You can use the command `@editorialbot generate pdf` to re-compile the pdf once you've updated the text.
