### 2023-01-23 jrzaurin/pytorch-widedeep#127 @jrzaurin
Hey @siboehm

I am not sure about A, you are right, that needs to be checked. Let me have a look. I am bit busy this week, but let's see if between me and @5uperpalo we find a sec.

Thanks!
