### 2023-01-30 jrzaurin/pytorch-widedeep#127 @5uperpalo
as I understand it both licenses can be combined but we need to change the license in the library from MIT to Apache 2.0 . many of the posts that I found refer to [this](https://web.archive.org/web/20210101030518/https://dwheeler.com/essays/floss-license-slide.html) post where the most permissive license is on the left and the common license that could be used to combine them is down the "stream of arrows".
* Also, according to more "restrictive"(brackets as in general there is tiny difference) Apache 2.0 license the MIT is similar, see [Apache documentation](https://apache.org/legal/resolved.html).
* to be on a "safe side" I think we should include both MIT and Apache 2.0 license as states in [this](https://law.stackexchange.com/a/6758) post:
```
Both the Apache License and the MIT license are permissive, so incorporating MIT licensed code into your Apache licensed project is certainly allowed. Just be sure to attribute the original author for the parts your incorporated and include a copy of the MIT License terms, as required by the license. Go ahead and read it to make sure you get all the details right; it's fairly short and easy for a layperson to comprehend.
```
* another [post](https://opensource.stackexchange.com/a/11654) agreeing with this statement; more profound [answer](https://www.quora.com/Is-the-MIT-license-compatible-with-the-Apache-License-Version-2-APLv2); [MIT vs Apache 2.0 compatibility](https://fossa.com/blog/open-source-licenses-101-mit-license/)

**Summary:**
To be on "a safest possible side" we include both licenses and we should credit authors of the code - which we did and continue doing.
