### 2023-01-30 7af79e2
added quantization option for continuous cols. Needs reviewing
