+++ pytorch_widedeep/preprocessing/tab_preprocessor.py
@@ -4,7 +4,14 @@ import numpy as np
 import pandas as pd
 from sklearn.preprocessing import StandardScaler
 
-from pytorch_widedeep.wdtypes import Dict, List, Tuple, Union, Literal
+from pytorch_widedeep.wdtypes import (
+    Dict,
+    List,
+    Tuple,
+    Union,
+    Literal,
+    Optional,
+)
 from pytorch_widedeep.utils.general_utils import Alias
 from pytorch_widedeep.utils.deeptabular_utils import LabelEncoder
 from pytorch_widedeep.preprocessing.base_preprocessor import (
@@ -34,6 +41,56 @@ def embed_sz_rule(
         return int(min(50, (n_cat // 2) + 1))
     else:
         return int(min(600, round(1.6 * n_cat**0.56)))
+
+
+class Quantizer:
+    def __init__(
+        self,
+        quantization_setup: Dict[str, Union[int, List[float]]],
+        **kwargs,
+    ):
+
+        self.quantization_setup = quantization_setup
+        self.quant_args = kwargs
+
+        self.is_fitted = False
+
+    def fit(self, df: pd.DataFrame) -> "Quantizer":
+
+        self.bins: Dict[str, List[float]] = {}
+        for col, bins in self.quantization_setup.items():
+            _, self.bins[col] = pd.cut(
+                df[col], bins, retbins=True, labels=False, **self.quant_args
+            )
+
+        self.inversed_bins: Dict[str, Dict[int, float]] = {}
+        for col, bins in self.bins.items():
+            self.inversed_bins[col] = {
+                k: v
+                for k, v in list(
+                    zip(
+                        range(len(bins)),
+                        [(a + b) / 2.0 for a, b in zip(bins, bins[1:])],
+                    )
+                )
+            }
+
+        self.is_fitted = True
+
+        return self
+
+    def transform(self, df: pd.DataFrame) -> pd.DataFrame:
+
+        check_is_fitted(self, condition=self.is_fitted)
+
+        dfc = df.copy()
+        for col, bins in self.bins.items():
+            dfc[col] = pd.cut(dfc[col], bins, labels=False, **self.quant_args)
+
+        return dfc
+
+    def fit_transform(self, df: pd.DataFrame) -> pd.DataFrame:
+        return self.fit(df).transform(df)
 
 
 class TabPreprocessor(BasePreprocessor):
@@ -145,6 +202,10 @@ class TabPreprocessor(BasePreprocessor):
         self,
         cat_embed_cols: Union[List[str], List[Tuple[str, int]]] = None,
         continuous_cols: List[str] = None,
+        quantization_setup: Optional[
+            Union[int, Dict[str, Union[int, List[float]]]]
+        ] = None,
+        cols_to_scale: Optional[List[str]] = None,
         scale: bool = True,
         already_standard: List[str] = None,
         auto_embed_dim: bool = True,
@@ -154,10 +215,13 @@ class TabPreprocessor(BasePreprocessor):
         with_cls_token: bool = False,
         shared_embed: bool = False,
         verbose: int = 1,
+        **kwargs,
     ):
         super(TabPreprocessor, self).__init__()
 
         self.continuous_cols = continuous_cols
+        self.quantization_setup = quantization_setup
+        self.cols_to_scale = cols_to_scale
         self.scale = scale
         self.already_standard = already_standard
         self.auto_embed_dim = auto_embed_dim
@@ -167,6 +231,13 @@ class TabPreprocessor(BasePreprocessor):
         self.with_cls_token = with_cls_token
         self.shared_embed = shared_embed
         self.verbose = verbose
+
+        self.quant_args = {
[…]
