### 2023-01-30 cce75e3
a few types fixed here and there. Still some work to do on typing
