+++ pytorch_widedeep/training/_base_bayesian_trainer.py
@@ -9,6 +9,7 @@ from torch.optim.lr_scheduler import ReduceLROnPlateau
 
 from pytorch_widedeep.metrics import Metric, MultipleMetrics
 from pytorch_widedeep.wdtypes import (
+    Any,
     List,
     Union,
     Module,
@@ -193,7 +194,10 @@ class BaseBayesianTrainer(ABC):
         else:
             self.cyclic_lr = False
 
-    def _set_callbacks_and_metrics(self, callbacks, metrics):
+    # this needs type fixing to adjust for the fact that the main class can
+    # take an 'object', a non-instastiated Class, so, should be something like:
+    # callbacks: Optional[List[Union[object, Callback]]] in all places
+    def _set_callbacks_and_metrics(self, callbacks: Any, metrics: Any):
         self.callbacks: List = [History(), LRShedulerCallback()]
         if callbacks is not None:
             for callback in callbacks:
