### 2023-02-02 7596ca4
Fixed a lot of typing issues. Adjusted tests. Needs review and update docs
