+++ pytorch_widedeep/models/wide_deep.py
@@ -1,27 +1,21 @@
-"""
-During the development of the package I realised that there is a typing
-inconsistency. The input components of a Wide and Deep model are of type
-nn.Module. These change type internally to nn.Sequential. While nn.Sequential
-is an instance of nn.Module the oppossite is, of course, not true. This does
-not affect any funcionality of the package, but it is something that needs
-fixing. However, while fixing is simple (simply define new attributes that
-are the nn.Sequential objects), its implications are quite wide within the
-package (involves changing a number of tests and tutorials). Therefore, I
-will introduce that fix when I do a major release. For now, we live with it.
-"""
-
 import warnings
 
 import torch
 from torch import nn
 
-from pytorch_widedeep.wdtypes import Dict, List, Tuple, Tensor, Optional
+from pytorch_widedeep.wdtypes import Dict, List, Tuple, Union, Tensor, Optional
 from pytorch_widedeep.models.fds_layer import FDSLayer
 from pytorch_widedeep.models._get_activation_fn import get_activation_fn
 from pytorch_widedeep.models.tabular.mlp._layers import MLP
 from pytorch_widedeep.models.tabular.tabnet.tab_net import TabNetPredLayer
+from pytorch_widedeep.models._base_wd_model_component import (
+    BaseWDModelComponent,
+)
 
 warnings.filterwarnings("default", category=UserWarning)
+
+
+WDModel = Union[nn.Module, nn.Sequential, BaseWDModelComponent]
 
 
 class WideDeep(nn.Module):
@@ -140,10 +134,10 @@ class WideDeep(nn.Module):
     def __init__(
         self,
         wide: Optional[nn.Module] = None,
-        deeptabular: Optional[nn.Module] = None,
-        deeptext: Optional[nn.Module] = None,
-        deepimage: Optional[nn.Module] = None,
-        deephead: Optional[nn.Module] = None,
+        deeptabular: Optional[BaseWDModelComponent] = None,
+        deeptext: Optional[BaseWDModelComponent] = None,
+        deepimage: Optional[BaseWDModelComponent] = None,
+        deephead: Optional[BaseWDModelComponent] = None,
         head_hidden_dims: Optional[List[int]] = None,
         head_activation: str = "relu",
         head_dropout: float = 0.1,
@@ -176,22 +170,16 @@ class WideDeep(nn.Module):
         # required as attribute just in case we pass a deephead
         self.pred_dim = pred_dim
 
-        # The main 5 components of the wide and deep assemble
-        self.wide = wide
-        self.deeptabular = deeptabular
-        self.deeptext = deeptext
-        self.deepimage = deepimage
-        self.deephead = deephead
-        self.enforce_positive = enforce_positive
         self.with_fds = with_fds
+        self.enforce_positive = enforce_positive
 
-        if self.deeptabular is not None:
-            self.is_tabnet = deeptabular.__class__.__name__ == "TabNet"
-        else:
-            self.is_tabnet = False
-
-        if self.deephead is None and head_hidden_dims is not None:
-            self._build_deephead(
+        # The main 5 components of the wide and deep assemble
+        self.with_deephead = deephead is not None or head_hidden_dims is not None
+        if deephead is None and head_hidden_dims is not None:
+            self.deephead = self._build_deephead(
+                deeptabular,
+                deeptext,
+                deepimage,
                 head_hidden_dims,
                 head_activation,
                 head_dropout,
@@ -199,12 +187,18 @@ class WideDeep(nn.Module):
                 head_batchnorm_last,
                 head_linear_first,
             )
-        elif self.deephead is not None:
+        elif deephead is not None:
             self.deephead = nn.Sequential(
[…]
