### 2023-02-19 #5027 @osorensen
> Went through it, the code itself mostly looks good but I have quite a few issues with the written text in it's current state.
>
> 1. Paper is missing a comparison to other similar packages in the field, as already pointed out by @makoeppel.
> 2. l55: In the formula a_deepimage seems to be missing after W_deepimage. The same formula in the documentation contains an a_deepimage term.
> 3. There are a lot of problems with wording and spelling in the current version of the paper. These should be fixed, before the paper can be accepted. The quickest way to resolve them is probably to pipe the whole text through Grammarly. Some issues (far from exhaustive!):
>
>    1. Pytorch -> PyTorch
>    2. l12: frames -> frameworks
>    3. l7: What's the different between a data type and a mode?
>    4. The sentence starting in l17 is either misspelled or unclear
>    5. The sentence starting in l32 is incomplete
>    6. l38: an -> a, plethora -> a plethora
> 4. In the acknowledgements section, the authors mention a lot of other sources of code. It's not clear to me whether the licences of these other projects have been respected (eg ZILNloss is licenced under Apache, but pytorch-widedeep is licensed under MIT, and there's no lincense info in losses.py). See [Licences of copied code jrzaurin/pytorch-widedeep#127](https://github.com/jrzaurin/pytorch-widedeep/issues/127)

@5uperpalo, could you please let us know how it's going addressing these points?
