+++ mkdocs/paper_JOSS/paper.md
@@ -24,11 +24,11 @@ bibliography: paper.bib
 
 # Summary
 
-In recent years datasets have grown both in size and diversity, combining different data types or modes. 
+In recent years datasets have grown both in size and diversity, combining different data types. 
 Multi-modal machine learning projects involving tabular data, images and/or text are gaining popularity (e.g. [@garg2022multimodality]).
-Traditional approaches involved independent feature generation from every data mode and their combination in later stage, before passing them to an algorithm for classification or regression.
+Traditional approaches involved independent feature generation from every data type and their combination in later stage, before passing them to an algorithm for classification or regression.
 
-However, with the advent of "_easy-to-use_" Deep Learning (DL) frames such as Tensorflow [@tensorflow2015-whitepaper] or Pytorch [@NEURIPS2019_9015], and the subsequent advances in the fields of Computer Vision, Natural Language Processing or Deep Learning for Tabular data, it is now possible to use SoTA DL models and combine all datasets early in the process. This has two main advantages: (i) we can partially, or entirely skip the feature engineering step and (ii) the representations of each data mode are learned jointly. This means that such representations contain information that not only relates to the target, if the problem is supervised, but to how the different data modes relates to each other.
+However, with the advent of "_easy-to-use_" Deep Learning (DL) frames such as Tensorflow [@tensorflow2015-whitepaper] or Pytorch [@NEURIPS2019_9015], and the subsequent advances in the fields of Computer Vision, Natural Language Processing or Deep Learning for Tabular data, it is now possible to use SoTA DL models and combine all datasets early in the process. This has two main advantages: (i) we can partially, or entirely skip the feature engineering step and (ii) the representations of each data type are learned jointly. This means that such representations contain information that not only relates to the target, if the problem is supervised, but to how the different data types relates to each other.
 
 Furthermore, the flexibility inherent to DL approaches allows the usage of techniques that are primarly designed only for the text and/or images, to tabular data, eg. transfer learning or self-supervised pre-training.
 
@@ -36,13 +36,13 @@ With that in mind, we introduce `pytorch-widedeep`, a flexible package for multi
 
 # Statement of need
 
-There is a small number of packages available to use DL for tabular data alone (e.g pytorch-tabular [@joseph2021pytorch], pytorch-tabnet or autogluon-tabular [@erickson2020autogluon]) or that focus mainly in combining text and images (e.g. MMF [@singh2020mmf]). With that in mind, our goal is to provide a modular, flexible, and "_easy-to-use_" frame that allows the combination of a wide variety of models for all data modes.
+There is a small number of packages available to use DL for tabular data alone (e.g pytorch-tabular [@joseph2021pytorch], pytorch-tabnet or autogluon-tabular [@erickson2020autogluon]) or that focus mainly in combining text and images (e.g. MMF [@singh2020mmf]). With that in mind, our goal is to provide a modular, flexible, and "_easy-to-use_" frame that allows the combination of a wide variety of models for all data types.
 
[…]
