+++ mkdocs/paper_JOSS/paper_bck.md
@@ -6,7 +6,7 @@ tags:
   - Deep learning
 authors:
   - name: Javier Rodriguez Zaurin
-    orcid: 0000-0000-0000-0000
+    orcid: 0000-0002-1082-1107
     # equal-contrib: true
     affiliation: 1 # (Multiple affiliations must be quoted)
   - name: Pavol Mulinka
@@ -18,29 +18,31 @@ affiliations:
    index: 1
  - name: Centre Tecnologic de Telecomunicacions de Catalunya (CTTC/CERCA), Catalunya, Spain
    index: 2
-date: 03 September 2022
+date: 13 November 2022
 bibliography: paper.bib
 ---
 
 # Summary
 
-In recent years datasets have grown both in size and diversity, combining different data types or modes. In fact, it is not unusual these days to face machine learning projects that involve tabular data, images and/or text. Traditionally, one would address these projects by independently generating some features from every data mode, combine them at a later stage and then pass them to an algorithm such a GBM to perform classification or regression (if, for example, the nature of the problem is supervised).
+In recent years datasets have grown both in size and diversity, combining different data types or modes. 
+Multi-modal machine learning projects involving tabular data, images and/or text are gaining popularity (e.g. [@garg2022multimodality]).
+Traditional approaches involved independent feature generation from every data mode and their combination in later stage, before passing them to an algorithm for classification or regression.
 
-However, with the advent of "_easy-to-use_" Deep Learning (DL) frames such as Tensorflow [@tensorflow2015-whitepaper] or Pytorch [@NEURIPS2019_9015], and the subsequent advances in the fields of Computer Vision, Natural Language Processing or Deep Learning for Tabular data, it is now possible to use SoTA DL models and combine all datasets earlier in the process. This has two main advantages: in the first place we can partially (or entirely) avoid the feature engineering step, which can be tedious in some cases. Secondly, the representations of each data mode, tabular, text and images, is learned jointly, harboring information that not only relates to the target, if the problem is supervised, but to how the different data modes relates to each other.
+However, with the advent of "_easy-to-use_" Deep Learning (DL) frames such as Tensorflow [@tensorflow2015-whitepaper] or Pytorch [@NEURIPS2019_9015], and the subsequent advances in the fields of Computer Vision, Natural Language Processing or Deep Learning for Tabular data, it is now possible to use SoTA DL models and combine all datasets early in the process. This has two main advantages: (i) we can partially, or entirely skip the feature engineering step and (ii) the representations of each data mode are learned jointly. This means that such representations contain information that not only relates to the target, if the problem is supervised, but to how the different data modes relates to each other.
 
-Furthermore, the flexibility inherent to using DL approaches allows the use of techniques that are principle designed for problems involving only text and/or images, to tabular data, such as for example transfer learning or self-supervised pre-training.
+Furthermore, the flexibility inherent to DL approaches allows the usage of techniques that are primarly designed only for the text and/or images, to tabular data, eg. transfer learning or self-supervised pre-training.
 
-With that in mind we introduce `pytorch-widedeep`, a flexible package for multimodal-deep-learning, designed to facilitate the combination of tabular data with text and images.
+With that in mind, we introduce `pytorch-widedeep`, a flexible package for multimodal-deep-learning, designed to facilitate the combination of tabular data with text and images.
 
 # Statement of need
 
[…]
