### 2023-02-23 jrzaurin/pytorch-widedeep#127 @5uperpalo
I put Apache2.0 and MIT license separately and mentioned it in the README.MD: https://github.com/jrzaurin/pytorch-widedeep/tree/joss_paper#license . Other than information from my previous post I also followed advice from here:
* https://softwareengineering.stackexchange.com/a/371435
* https://softwareengineering.stackexchange.com/a/371526
* LIcense Syntax https://en.wikipedia.org/wiki/Software_Package_Data_Exchange:
```
For example, (Apache-2.0 OR MIT) means that one can choose between Apache-2.0 (Apache License) or MIT (MIT license). On the other hand, (Apache-2.0 AND MIT) means that both licenses apply
```
I put the licenses in joss_paper branch and I will pull request to the main branch also
