### 2023-02-23 #5027 @5uperpalo
first of all, thank you all for the precious time that you are spending on this... my reaction to the previous comments:
1. "Paper is missing a comparison..." I hope we resolved this in this related issue: https://github.com/jrzaurin/pytorch-widedeep/issues/124
2. "l55: In the formula a_deepimage seems to be missing after W_deepimage." - thank you for noticing, I fixed it
3. "There are a lot of problems with wording and spelling... " thank you for noticing I put it through Grammarly as you suggested + re-read it + fixed minor issues (data mode/type - same meaning, I just used type as is more usual, frame/framework, etc)
4. I put both MIT and APACHE2.0 licenses in the repo and mentioned it in the README.MD and I am doing the same for the main branch; see the links and logic I followed for this licensing issue https://github.com/jrzaurin/pytorch-widedeep/issues/127

@makoeppel @osorensen @siboehm @dataplayer12 I do firmly believe we should/could be approaching the final stage 😃
