### 2023-03-17 cc0d1ad
unit tested the col quantization. Fixed some types and adjusted to new version of torchvision
