### 2023-03-17 775de22
set scale to True in the examples
