+++ examples/scripts/adult_census_tabnet.py
@@ -43,6 +43,7 @@ if __name__ == "__main__":
     tab_preprocessor = TabPreprocessor(
         cat_embed_cols=cat_embed_cols,
         continuous_cols=continuous_cols,
+        scale=True,
         default_embed_dim=1,
     )
     X_tab = tab_preprocessor.fit_transform(df)
