### 2023-03-22 46bf299
Adjusted wdtypes to the fact that we do not support 3.7 anymore
