### 2023-03-29 jrzaurin/pytorch-widedeep#135 @osorensen
Multimodal deep learning
This is an editorial comment to your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/5027).

The title of the paper is "pytorch-widedeep: A flexible package for multimodal-deep-learning". To the best of my knowledge, a more correct spelling would be "pytorch-widedeep: A flexible package for multimodal deep learning". Please update and report here when done.
