### 2023-03-29 jrzaurin/pytorch-widedeep#136 @osorensen
Reference formatting
This is an editorial comment to your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/5027).

Please change the reference so it appears as "(e.g., Garg et al. (2022))". This can be done by changing the line in markdown from "(e.g. [@garg2022multimodality])" to "(e.g. @garg2022multimodality)".
