### 2023-03-29 jrzaurin/pytorch-widedeep#137 @osorensen
SoTA DL?
This is an editorial comment to your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/5027).

On line 15 you mention SoTA DL models. Please introduce the abbreviation SoTA before its first use.
