### 2023-03-29 jrzaurin/pytorch-widedeep#138 @osorensen
Minor grammatical thing
This is an editorial comment to your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/5027).

On line 21, I think it makes sense to remove "the" in front of "text", as highlighted below:
