### 2023-03-29 jrzaurin/pytorch-widedeep#139 @osorensen
Inconsistent spelling
Editorial comment to your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/5027).

On line 8 you write "multi-modal". Please be consistent and change to "multimodal".
