### 2023-03-29 jrzaurin/pytorch-widedeep#135 @osorensen
Please also change from "multimodel-deep-learning" to "multimodal deep learning" on lines 23-24.
