### 2023-03-29 jrzaurin/pytorch-widedeep#140 @osorensen
Minor grammatical thing
This is an editorial comment to your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/5027).

On line 28, please change from "focus mainly in" to "focus mainly on".
