### 2023-03-29 jrzaurin/pytorch-widedeep#141 @osorensen
Reference to Figure 1
This is an editorial comment to your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/5027).

On line 41, please make sure the word "Figure" appears only once:
