### 2023-03-29 jrzaurin/pytorch-widedeep#142 @osorensen
Quotes
This is an editorial comment to your https://github.com/openjournals/joss-reviews/issues/5027.

On line 45, please remove the space between the quotes " and 'on top':
