### 2023-03-29 jrzaurin/pytorch-widedeep#143 @osorensen
Equation formatting
This is an editorial comment on your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/5027).

The equations on lines 54-55 and on line 64 do not appear correctly. Please follow to [author guidelines](https://joss.readthedocs.io/en/latest/submitting.html) and use `$$` for equations that appear on a single line, or LaTeX:

> # Mathematics
>
> Single dollars ($) are required for inline mathematics e.g. $f(x) = e^{\pi/x}$
>
> Double dollars make self-standing equations:
>
> $$\Theta(x) = \left\{\begin{array}{l}
>0\textrm{ if } x < 0\cr
> 1\textrm{ else}
> \end{array}\right.$$
>
> You can also use plain \LaTeX for equations
> \begin{equation}\label{eq:fourier}
> \hat f(\omega) = \int_{-\infty}^{\infty} f(x) e^{i\omega x} dx
> \end{equation}
> and refer to \autoref{eq:fourier} from text.
