### 2023-03-29 jrzaurin/pytorch-widedeep#144 @osorensen
Cross product transformation
This is an editorial comment to your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/5027).

Please remove the following part from the paper, lines 58-62:

> In case you are wondering what are _"cross product transformations"_, here is a quote taken directly from the paper: _"For binary features, a cross-product transformation (e.g., "AND(gender=female, language=en)") is 1 if and only if the constituent features ("gender=female" and "language=en") are all 1, and 0 otherwise"._
