### 2023-03-29 jrzaurin/pytorch-widedeep#145 @osorensen
Typo
This is an editorial comment to your https://github.com/openjournals/joss-reviews/issues/5027.

Please remove the dash at the end of a sentence on line 74, highlighted below:
