### 2023-03-29 jrzaurin/pytorch-widedeep#146 @osorensen
Capitalization
This is an editorial comment to your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/5027).

On line 91, the word "Weight" should probably be changed to "weight"?
