### 2023-03-29 jrzaurin/pytorch-widedeep#147 @osorensen
Missing '@' in reference
This is an editorial comment to your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/5027).

Please fix the reference to Efficientnet on line 101:
