### 2023-03-29 jrzaurin/pytorch-widedeep#147 @osorensen
Also fix a similar issue on line 106:
