### 2023-03-29 jrzaurin/pytorch-widedeep#148 @osorensen
Slack
This is an editorial comment to your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/5027).

On line 116, please change from "slack" to "Slack", with capital S.
