### 2023-03-29 jrzaurin/pytorch-widedeep#148 @osorensen
Also please remove the reference to Pytorch-Widedeep Slack.
