### 2023-03-29 jrzaurin/pytorch-widedeep#149 @osorensen
Keras citation
This is an editorial comment to your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/5027).

Please update citation so Keras is cited properly, see https://keras.io/getting_started/faq/#how-should-i-cite-keras.
