+++ mkdocs/paper_JOSS/paper.bib
@@ -341,11 +341,12 @@ doi = {10.1162/neco.1997.9.8.1735}
   note="[Online]. Available at: https://github.com/ncullen93/torchsample"
 }
 
-@manual{keras,
-  doi={10.1163/1574-9347_dnp_e612900},
-  title={keras},
-  note="[Online]. Available at: https://keras.io/"
-}
+@misc{chollet2015keras,
+  title={Keras},
+  author={Chollet, Fran\c{c}ois and others},
+  year={2015},
+  howpublished={\url{https://keras.io}},
+} 
 
 @manual{fastai_tokenizer,
   title={fastai tokenizer},
