### 2023-04-19 jrzaurin/pytorch-widedeep#137 @5uperpalo
fixed in [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05027/joss.05027/10.21105.joss.05027.pdf)
![image](https://user-images.githubusercontent.com/36483359/233193752-b11b194c-c4ff-49b8-8e94-d944591ee768.png)
I followed this advice to use it with hyphens as adjective: https://writingexplained.org/state-of-the-art-vs-state-of-the-art
