### 2023-04-19 jrzaurin/pytorch-widedeep#138 @5uperpalo
fixed in [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05027/joss.05027/10.21105.joss.05027.pdf)
![image](https://user-images.githubusercontent.com/36483359/233194060-f28dc32c-89c9-4dd2-89ec-e87e6be0f9fe.png)
