### 2023-04-19 jrzaurin/pytorch-widedeep#139 @5uperpalo
fixed in [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05027/joss.05027/10.21105.joss.05027.pdf)
![image](https://user-images.githubusercontent.com/36483359/233194330-7e4f91f6-52a8-4064-b8cf-a7f519e391ce.png)

![image](https://user-images.githubusercontent.com/36483359/233194394-27a161d2-b19a-4ed4-b413-eb2f341f49be.png)

![image](https://user-images.githubusercontent.com/36483359/233194612-51186b92-f0ae-45f9-930f-c2227df3de2c.png)
