### 2023-04-19 jrzaurin/pytorch-widedeep#141 @5uperpalo
fixed in [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05027/joss.05027/10.21105.joss.05027.pdf)
![image](https://user-images.githubusercontent.com/36483359/233197013-22a982a2-ba81-47de-9cb3-87ececa19c0f.png)
