### 2023-04-19 jrzaurin/pytorch-widedeep#142 @5uperpalo
fixed in [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05027/joss.05027/10.21105.joss.05027.pdf)
![image](https://user-images.githubusercontent.com/36483359/233198087-bc037f1a-c27a-47b3-97ac-85a5dc436c79.png)
