### 2023-04-19 jrzaurin/pytorch-widedeep#143 @5uperpalo
fixed in [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05027/joss.05027/10.21105.joss.05027.pdf)
![image](https://user-images.githubusercontent.com/36483359/233198305-a1b0ab10-dcac-49a6-bbbe-a0a1762b1648.png)

![image](https://user-images.githubusercontent.com/36483359/233198512-932c8db2-2270-4355-89fc-4d6f9fd68deb.png)
