### 2023-04-19 jrzaurin/pytorch-widedeep#144 @5uperpalo
fixed in [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05027/joss.05027/10.21105.joss.05027.pdf)

![image](https://user-images.githubusercontent.com/36483359/233198881-4dff77d2-829b-477d-acab-952087ddd3be.png)
