### 2023-04-19 jrzaurin/pytorch-widedeep#146 @5uperpalo
fixed in [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05027/joss.05027/10.21105.joss.05027.pdf)
The Weight uncertainty in the neural network is(was) title of the paper, I kept only the reference, see:
![image](https://user-images.githubusercontent.com/36483359/233199424-5e85d3b2-f3e2-4ddc-9e78-b18b717b1402.png)
