### 2023-04-19 jrzaurin/pytorch-widedeep#148 @5uperpalo
fixed in [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05027/joss.05027/10.21105.joss.05027.pdf)
![image](https://user-images.githubusercontent.com/36483359/233200533-f30bf305-1d96-4e79-b58f-e2a98b6229cc.png)
