### 2023-04-19 jrzaurin/pytorch-widedeep#149 @5uperpalo
fixed in [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05027/joss.05027/10.21105.joss.05027.pdf)
![image](https://user-images.githubusercontent.com/36483359/233200797-0463f59d-7953-4a68-880a-5f57e2c50377.png)

![image](https://user-images.githubusercontent.com/36483359/233201018-238109cb-1488-4536-b5a9-7f5742c5cfe9.png)
