### 2023-04-19 jrzaurin/pytorch-widedeep#147 @5uperpalo
fixed in [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05027/joss.05027/10.21105.joss.05027.pdf)
![image](https://user-images.githubusercontent.com/36483359/233202091-87e7d6ca-c17a-4fa8-8791-bbccfec951a8.png)

![image](https://user-images.githubusercontent.com/36483359/233202185-79da4cbb-9d7f-43d4-bf31-f9e775b5175f.png)
