### 2023-04-19 #5027 @5uperpalo
@osorensen thank you for your comments, I fixed all the issues you pointed out, I am also sorry for the delay, but it was a holiday season - which means a lot of additional work while colleagues are gone for vacations :) :P ; please look at the the fixes and let me know if you have any other comments
