### 2023-04-21 6c2b9f7
Added a class to compute feature importance. I need to extend it to an explain capability and test the hell out of it
