+++ examples/scripts/adult_census_feature_importance.py
@@ -1,13 +1,15 @@
-import torch
 from sklearn.model_selection import train_test_split
 
 from pytorch_widedeep import Trainer
 from pytorch_widedeep.models import (
     SAINT,
+    TabNet,
     WideDeep,
     FTTransformer,
     TabFastFormer,
     TabTransformer,
+    SelfAttentionMLP,
+    ContextAttentionMLP,
 )
 from pytorch_widedeep.metrics import Accuracy
 from pytorch_widedeep.datasets import load_adult
@@ -19,20 +21,25 @@ df = load_adult(as_frame=True)
 df.columns = [c.replace("-", "_") for c in df.columns]
 df["income_label"] = (df["income"].apply(lambda x: ">50K" in x)).astype(int)
 df.drop("income", axis=1, inplace=True)
-target = "income_label"
+target_colname = "income_label"
 
 cat_embed_cols = []
 for col in df.columns:
-    if df[col].dtype == "O" or df[col].nunique() < 200 and col != target:
+    if df[col].dtype == "O" or df[col].nunique() < 200 and col != target_colname:
         cat_embed_cols.append(col)
 
-train, test = train_test_split(df, test_size=0.1, random_state=1, stratify=df[[target]])
+train, test = train_test_split(
+    df, test_size=0.1, random_state=1, stratify=df[[target_colname]]
+)
 
-tab_preprocessor = TabPreprocessor(cat_embed_cols=cat_embed_cols, with_attention=True)
+with_cls_token = True
+tab_preprocessor = TabPreprocessor(
+    cat_embed_cols=cat_embed_cols, with_attention=True, with_cls_token=with_cls_token
+)
 
 X_tab_train = tab_preprocessor.fit_transform(train)
 X_tab_test = tab_preprocessor.transform(test)
-target = train[target].values
+target = train[target_colname].values
 
 tab_transformer = TabTransformer(
     column_idx=tab_preprocessor.column_idx,
@@ -47,7 +54,7 @@ saint = SAINT(
     cat_embed_input=tab_preprocessor.cat_embed_input,
     input_dim=8,
     n_heads=2,
-    n_blocks=1,
+    n_blocks=2,
 )
 
 tab_fastformer = TabFastFormer(
@@ -55,7 +62,7 @@ tab_fastformer = TabFastFormer(
     cat_embed_input=tab_preprocessor.cat_embed_input,
     input_dim=8,
     n_heads=2,
-    n_blocks=1,
+    n_blocks=2,
 )
 
 ft_transformer = FTTransformer(
@@ -64,11 +71,69 @@ ft_transformer = FTTransformer(
     input_dim=8,
     n_heads=2,
     n_blocks=2,
-    kv_compression_factor=1.0,
+    kv_compression_factor=1.0,  # if this is diff than one, we cannot do this
 )
 
+context_attentio_mlp = ContextAttentionMLP(
+    column_idx=tab_preprocessor.column_idx,
+    cat_embed_input=tab_preprocessor.cat_embed_input,
+    input_dim=16,
+    attn_dropout=0.2,
+    n_blocks=3,
+)
 
-model = WideDeep(deeptabular=tab_transformer)
+self_attention_mlp = SelfAttentionMLP(
+    column_idx=tab_preprocessor.column_idx,
+    cat_embed_input=tab_preprocessor.cat_embed_input,
+    n_blocks=3,
+)
+
+for attention_based_model in [
+    tab_transformer,
+    saint,
+    tab_fastformer,
+    ft_transformer,
+    context_attentio_mlp,
+    self_attention_mlp,
+]:
+    model = WideDeep(deeptabular=attention_based_model)  # type: ignore[arg-type]
+
[…]

[…]
