+++ pytorch_widedeep/training/trainer.py
@@ -1204,6 +1204,7 @@ class Trainer(BaseTrainer):
             "generator",
             "prefetch_factor",
             "persistent_workers",
+            "oversample_mul",
         ]
         lds_params = [
             "lds_kernel",
