### 2023-04-25 124fdb1
added oversample_mul to the dataloader_params to (manually) make it consistent with master
