### 2023-04-25 jrzaurin/pytorch-widedeep#156 @osorensen
Editorial comments in JOSS submission
Dear @5uperpalo

Thanks for addressing all the issues I raised about the paper. I have a few more based on a second reading of the revised paper, and this is the first one.

On lines 32-33, I suggest to remove the sentence underlined in red in this screenshot, because this is not very important information in a short paper describing the library:

https://github.com/openjournals/joss-reviews/issues/5027
