### 2023-04-25 jrzaurin/pytorch-widedeep#157 @osorensen
Editorial comment
On line 36, please add a space between "Deep and Cross" and the parenthesized reference:

https://github.com/openjournals/joss-reviews/issues/5027
