### 2023-04-25 jrzaurin/pytorch-widedeep#158 @osorensen
Editorial comment
Please add a small caption to Figure 1, briefly describing what it shows.

You might want to move one or more sentences from the main text to the caption. Alternatively, you could simply write something like "Figure 1. Main components of the pytorch-widedeep architecture."

https://github.com/openjournals/joss-reviews/issues/5027
