### 2023-04-25 jrzaurin/pytorch-widedeep#159 @osorensen
Editorial comment
I suggest rewriting the highlighted text to "Following the notation of Cheng et al. (2016), ..."

https://github.com/openjournals/joss-reviews/issues/5027
