### 2023-04-25 jrzaurin/pytorch-widedeep#160 @osorensen
Editorial comment
This equation is too wide for a single line.

I suggest to write it more compact using summation, as

$$pred = \sigma(W_{wide}^{T}[x,\phi(x)] + \sum_{i \in \mathcal{I}} W_{i}^{T}a_{i}^{l_f} + b)$$

where you then need to add a description in the text below that `\mathcal{I} = \{deeptabular, deeptext, deepimage \}`.

The LaTeX for this is

```
pred = \sigma(W_{wide}^{T}[x,\phi(x)] + \sum_{i \in \mathcal{I}} W_{i}^{T}a_{i}^{l_f} + b)
```

https://github.com/openjournals/joss-reviews/issues/5027
