### 2023-04-25 jrzaurin/pytorch-widedeep#161 @osorensen
Editorial comment
Please remove "While" at the beginning of line 57, so it becomes "If there is a ...":

https://github.com/openjournals/joss-reviews/issues/5027
