### 2023-04-25 jrzaurin/pytorch-widedeep#163 @osorensen
TorchVision citations
Please remove the separate citations to TorchVision Models and TorchVision multi weight support, and instead cite TorchVision [as they request](https://pypi.org/project/torchvision/)

```
@software{torchvision2016,
    title        = {TorchVision: PyTorch's Computer Vision library},
    author       = {TorchVision maintainers and contributors},
    year         = 2016,
    journal      = {GitHub repository},
    publisher    = {GitHub},
    howpublished = {\url{https://github.com/pytorch/vision}}
}
```

https://github.com/openjournals/joss-reviews/issues/5027
