### 2023-04-25 jrzaurin/pytorch-widedeep#165 @osorensen
Fastai tokenizer
The citation to Fastai tokenizer is not very informative. I think this paper is a better reference to fastai: https://www.mdpi.com/2078-2489/11/2/108

https://github.com/openjournals/joss-reviews/issues/5027
