### 2023-04-25 jrzaurin/pytorch-widedeep#166 @osorensen
Deep learning for computer vision
Please cite the book by Adrian Rosebrock using a proper BibTeX entry

Something like this should get you started:

```
@Book{citekey,
  author    = "",
  title     = "",
  publisher = "",
  year      = ""
}
```

https://github.com/openjournals/joss-reviews/issues/5027
