+++ mkdocs/paper_JOSS/paper.bib
@@ -326,20 +326,24 @@ doi = {10.1162/neco.1997.9.8.1735}
   note="[Online]. Available at: https://github.com/jrzaurin/pytorch-widedeep/tree/master/examples"
 }
 
-@manual{torchvision_models,
-  title={torchvision models},
-  note="[Online]. Available at: https://pytorch.org/vision/stable/models.html"
+@software{torchvision2016,
+    title        = {TorchVision: PyTorch's Computer Vision library},
+    author       = {TorchVision maintainers and contributors},
+    year         = 2016,
+    journal      = {GitHub repository},
+    publisher    = {GitHub},
+    howpublished = {\url{https://github.com/pytorch/vision}}
 }
 
-@manual{torchvision_weight,
-  title={torchvision multi weight support},
-  note="[Online]. Available at: https://pytorch.org/blog/introducing-torchvision-new-multi-weight-support-api/"
+@software{torch_sample,
+    title        = {TorchSample: Lightweight pytorch functions for neural network featuremap sampling},
+    author       = {TorchSample maintainers and contributors},
+    year         = 2017,
+    journal      = {GitHub repository},
+    publisher    = {GitHub},
+    howpublished = {\url{https://github.com/ncullen93/torchsample}}
 }
 
-@manual{torch_sample,
-  title={torch sample},
-  note="[Online]. Available at: https://github.com/ncullen93/torchsample"
-}
 
 @misc{chollet2015keras,
   title={Keras},
@@ -348,14 +352,25 @@ doi = {10.1162/neco.1997.9.8.1735}
   howpublished={\url{https://keras.io}},
 } 
 
-@manual{fastai_tokenizer,
-  title={fastai tokenizer},
-  note="[Online]. Available at: https://docs.fast.ai/text.transform.html#BaseTokenizer.tokenizer"
+@Article{info11020108,
+  AUTHOR = {Howard, Jeremy and Gugger, Sylvain},
+  TITLE = {Fastai: A Layered API for Deep Learning},
+  JOURNAL = {Information},
+  VOLUME = {11},
+  YEAR = {2020},
+  NUMBER = {2},
+  ARTICLE-NUMBER = {108},
+  URL = {https://www.mdpi.com/2078-2489/11/2/108},
+  ISSN = {2078-2489},
+  ABSTRACT = {fastai is a deep learning library which provides practitioners with high-level components that can quickly and easily provide state-of-the-art results in standard deep learning domains, and provides researchers with low-level components that can be mixed and matched to build new approaches. It aims to do both things without substantial compromises in ease of use, flexibility, or performance. This is possible thanks to a carefully layered architecture, which expresses common underlying patterns of many deep learning and data processing techniques in terms of decoupled abstractions. These abstractions can be expressed concisely and clearly by leveraging the dynamism of the underlying Python language and the flexibility of the PyTorch library. fastai includes: a new type dispatch system for Python along with a semantic type hierarchy for tensors; a GPU-optimized computer vision library which can be extended in pure Python; an optimizer which refactors out the common functionality of modern optimizers into two basic pieces, allowing optimization algorithms to be implemented in 4&ndash;5 lines of code; a novel 2-way callback system that can access any part of the data, model, or optimizer and change it at any point during training; a new data block API; and much more. We used this library to successfully create a complete deep learning course, which we were able to write more quickly than using previous approaches, and the code was more clear. The library is already in wide use in research, industry, and teaching.},
+  DOI = {10.3390/info11020108}
 }
 
-@manual{dl4cv,
-  title={dl4cv},
[…]
