### 2023-04-25 jrzaurin/pytorch-widedeep#156 @5uperpalo
thank you for the comment, the issue is resolved in the latest version of the [paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05027/joss.05027/10.21105.joss.05027.pdf).
![image](https://user-images.githubusercontent.com/36483359/234399071-b0350d91-291e-43ee-977e-8baa4b31cbf3.png)
