### 2023-04-25 jrzaurin/pytorch-widedeep#157 @5uperpalo
thank you for the comment, the issue is resolved in the latest version of the [paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05027/joss.05027/10.21105.joss.05027.pdf).
![image](https://user-images.githubusercontent.com/36483359/234399573-2dc6ec10-115a-4015-a66b-e80bcdf0d15e.png)
