### 2023-04-25 jrzaurin/pytorch-widedeep#158 @5uperpalo
thank you for the comment, the issue is resolved in the latest version of the [paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05027/joss.05027/10.21105.joss.05027.pdf). I moved all relevant picture information into caption.
![image](https://user-images.githubusercontent.com/36483359/234399790-b6b69e5e-ac67-4a48-9f50-421e8c9212c6.png)
