### 2023-04-25 jrzaurin/pytorch-widedeep#159 @5uperpalo
thank you for the comment, the issue is resolved in the latest version of the [paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05027/joss.05027/10.21105.joss.05027.pdf).
![image](https://user-images.githubusercontent.com/36483359/234400510-6d0a9a2a-cbb0-46d9-948e-7205534d7b09.png)
