### 2023-04-25 jrzaurin/pytorch-widedeep#161 @5uperpalo
thank you for the comment, the issue is resolved in the latest version of the [paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05027/joss.05027/10.21105.joss.05027.pdf).
![image](https://user-images.githubusercontent.com/36483359/234400756-20a0edb9-5019-4487-bc66-0d2368ae141d.png)
