### 2023-04-25 jrzaurin/pytorch-widedeep#163 @5uperpalo
thank you for the comment, the issue is resolved in the latest version of the [paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05027/joss.05027/10.21105.joss.05027.pdf).

![image](https://user-images.githubusercontent.com/36483359/234401345-d08c2bda-7cdc-458a-8fa9-698dd6c21e8d.png)

![image](https://user-images.githubusercontent.com/36483359/234401430-6d6e059f-8a62-491a-b1b1-1965bf2029ee.png)
