### 2023-04-25 jrzaurin/pytorch-widedeep#165 @5uperpalo
thank you for the comment, the issue is resolved in the latest version of the [paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05027/joss.05027/10.21105.joss.05027.pdf).

![image](https://user-images.githubusercontent.com/36483359/234401981-c2b5fff7-a0f8-4a3d-bd32-c3bdb10f7ac6.png)

![image](https://user-images.githubusercontent.com/36483359/234402046-b879d50f-8fbf-40e2-aa65-4b46da11ba7b.png)
