### 2023-04-25 jrzaurin/pytorch-widedeep#166 @5uperpalo
thank you for the comment, the issue is resolved in the latest version of the [paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05027/joss.05027/10.21105.joss.05027.pdf).

![image](https://user-images.githubusercontent.com/36483359/234402226-31a19a43-c2fa-4d75-8c47-06fef430b3c9.png)

![image](https://user-images.githubusercontent.com/36483359/234402299-5224184f-a6b7-4184-818d-2da0a1b71eab.png)
