+++ pytorch_widedeep/training/trainer.py
@@ -8,8 +8,6 @@ import torch
 import torch.nn.functional as F
 from tqdm import trange
 from torch import nn
-
-# from scipy.sparse import csc_matrix
 from torchmetrics import Metric as TorchMetric
 from torch.utils.data import DataLoader
 
@@ -40,7 +38,10 @@ from pytorch_widedeep.training._trainer_utils import (
     wd_train_val_split,
     print_loss_and_metric,
 )
-from pytorch_widedeep.training._feature_importance import FeatureImportance
+from pytorch_widedeep.training._feature_importance import (
+    Explainer,
+    FeatureImportance,
+)
 
 
 class Trainer(BaseTrainer):
@@ -532,11 +533,9 @@ class Trainer(BaseTrainer):
         self.callback_container.on_train_end(epoch_logs)
 
         if feature_importance_sample_size is not None:
-            self.feature_importance = FeatureImportance(self.device).feature_importance(
-                train_loader, self.model
-            )
-        # if self.model.is_tabnet:
-        #     self._compute_feature_importance(train_loader)
+            self.feature_importance = FeatureImportance(
+                self.device, feature_importance_sample_size
+            ).feature_importance(train_loader, self.model)
         self._restore_best_weights()
         self.model.train()
 
@@ -748,67 +747,16 @@ class Trainer(BaseTrainer):
         if self.method == "multiclass":
             return np.vstack(preds_l)
 
-    # def explain(
-    #     self, X_tab: np.ndarray, save_step_masks: bool = False
-    # ) -> Union[Tuple, np.ndarray]:
-    #     """
-    #     if the `deeptabular` component is a `Tabnet` model, returns the
-    #     aggregated feature importance for each instance (or observation) in
-    #     the `X_tab` array. If `save_step_masks` is set to `True`, the
-    #     masks per step will also be returned.
+    def explain(self, X_tab: np.ndarray, save_step_masks: Optional[bool] = None):
+        # TO DO: Add docs to this, to the feat imp parameter and the all
+        # related classes
+        model_explainer = Explainer(self.device)
 
-    #     Parameters
-    #     ----------
-    #     X_tab: np.ndarray
-    #         Input array corresponding **only** to the deeptabular component
-    #     save_step_masks: bool
-    #         Boolean indicating if the masks per step will be returned
+        res = model_explainer.explain(
+            self.model, X_tab, self.num_workers, self.batch_size, save_step_masks
+        )
 
-    #     Returns
-    #     -------
-    #     Union[Tuple, np.ndarray]
-    #         Array or Tuple of two arrays with the corresponding aggregated
-    #         feature importance and the masks per step if `save_step_masks`
-    #         is set to `True`
-    #     """
-    #     loader = DataLoader(
-    #         dataset=WideDeepDataset(X_tab=X_tab),
-    #         batch_size=self.batch_size,
-    #         num_workers=self.num_workers,
-    #         shuffle=False,
-    #     )
-
-    #     self.model.eval()
-    #     tabnet_backbone = list(self.model.deeptabular.children())[0]
-
-    #     m_explain_l = []
-    #     for batch_nb, data in enumerate(loader):
-    #         X = data["deeptabular"].to(self.device)
-    #         M_explain, masks = tabnet_backbone.forward_masks(X)  # type: ignore[operator]
-    #         m_explain_l.append(
-    #             csc_matrix.dot(M_explain.cpu().detach().numpy(), self.reducing_matrix)
-    #         )
-    #         if save_step_masks:
-    #             for key, value in masks.items():
-    #                 masks[key] = csc_matrix.dot(
[…]
