+++ pytorch_widedeep/training/_feature_importance.py
@@ -1,12 +1,23 @@
-from typing import Tuple, Union, Optional
+from abc import ABC, abstractmethod
 
 import numpy as np
 import torch
 from scipy.sparse import csc_matrix
 from torch.utils.data import DataLoader
 
+from pytorch_widedeep.wdtypes import (
+    Any,
+    Dict,
+    List,
+    Tuple,
+    Union,
+    Tensor,
+    Optional,
+    WideDeep,
+)
 from pytorch_widedeep.models.tabular import (
     SAINT,
+    TabPerceiver,
     FTTransformer,
     TabFastFormer,
     TabTransformer,
@@ -17,40 +28,98 @@ from pytorch_widedeep.utils.general_utils import Alias
 from pytorch_widedeep.training._wd_dataset import WideDeepDataset
 from pytorch_widedeep.models.tabular.tabnet._utils import create_explain_matrix
 
+TransformerBasedModels = (
+    SAINT,
+    FTTransformer,
+    TabFastFormer,
+    TabTransformer,
+    SelfAttentionMLP,
+    ContextAttentionMLP,
+)
+
+__all__ = ["FeatureImportance", "Explainer"]
+
+
+# TODO: review typing for the attention weights.
+# TODO: review typing for WideDeep (in particular the deeptabular part)
+
 
 class FeatureImportance:
-    def __init__(self, device, n_samples=1000):
+    def __init__(self, device: str, n_samples: int = 1000):
         self.device = device
         self.n_samples = n_samples
 
-    def feature_importance(self, loader, model):
-        self._check_inputs(model)
-        self.model_type = self._model_type(model)
+    def feature_importance(
+        self, loader: DataLoader, model: WideDeep
+    ) -> Dict[str, float]:
+        if model.is_tabnet:
+            model_feat_importance: ModelFeatureImportance = TabNetFeatureImportance(
+                self.device, self.n_samples
+            )
+        elif isinstance(model.deeptabular._modules["0"], TransformerBasedModels):
+            model_feat_importance = TransformerBasedFeatureImportance(
+                self.device, self.n_samples
+            )
+        else:
+            raise ValueError(
+                "The computation of feature importance is not supported for this particular model"
+            )
 
-        X = self._sample_data(loader)
+        return model_feat_importance.feature_importance(loader, model)
 
-        model.eval()
-        _ = model.deeptabular(X)
 
-        attention_weights = self._model_type_attention_weights(model)
+class Explainer:
+    def __init__(self, device: str):
+        self.device = device
 
-        feature_importance, column_idx = self._model_type_feature_importance(
-            model, attention_weights
-        )
-
-        return {k: v for k, v in zip(column_idx, feature_importance)}
-
-    @Alias("X_tab", "X")
     def explain(
         self,
-        model,
-        X_tab,
+        model: WideDeep,
+        X_tab: np.ndarray,
         num_workers: int,
         batch_size: Optional[int] = None,
-    ):
-        pass
+        save_step_masks: Optional[bool] = None,
+    ) -> Union[Tuple, np.ndarray]:
+        if model == "tabnet":
+            assert save_step_masks is not None
+            model_explainer: ModelExplainer = TabNetExplainer(self.device)
+            res = model_explainer.explain(
+                model,
+                X_tab,
+                num_workers,
+                batch_size,
+                save_step_masks,
+            )
+        elif isinstance(model.deeptabular._modules["0"], TransformerBasedModels):
+            model_explainer = TransformerBasedExplainer(self.device)
+            res = model_explainer.explain(
+                model,
+                X_tab,
+                num_workers,
+                batch_size,
+            )
+        else:
[…]
