### 2023-05-03 0d4a2fa
minor fix of attention weights typing
