+++ pytorch_widedeep/models/tabular/mlp/context_attention_mlp.py
@@ -188,7 +188,7 @@ class ContextAttentionMLP(BaseTabularModelWithAttention):
         )
 
     @property
-    def attention_weights(self) -> List:
+    def attention_weights(self) -> List[Tensor]:
         r"""List with the attention weights per block
 
         The shape of the attention weights is $(N, F)$, where $N$ is the batch
