+++ pytorch_widedeep/models/tabular/transformers/ft_transformer.py
@@ -275,7 +275,7 @@ class FTTransformer(BaseTabularModelWithAttention):
         )
 
     @property
-    def attention_weights(self) -> List:
+    def attention_weights(self) -> List[Tensor]:
         r"""List with the attention weights per block
 
         The shape of the attention weights is: $(N, H, F, k)$, where $N$ is
