### 2023-05-08 #5027 @osorensen
Thanks @jrzaurin. Can you please change the title of the Zenodo archive so it matches the paper title exactly? That is, change from "jrzaurin/pytorch-widedeep: Zenodo Release" to "pytorch-widedeep: A flexible package for multimodal-deep-learning".
