### 2023-05-08 #5027 @jrzaurin
hey @osorensen. Thanks for your answer.

The title of the release has been updated and I guess this will sync with the Zenodo archive. Otherwise we will have to publish a new release (?)

(Cc: @5uperpalo )
