### 2023-05-18 #5027 @jrzaurin
Hey, I tried to simply create a new release and I got a citations error:

Whenever I have a sec we will try to fix it
