+++ CITATION.cff
@@ -1,17 +1,0 @@
-cff-version: 1.2.0
-message: "If you want to cite the framework, feel free to use this"
-title: "pytorch-widedeep: A flexible package for multimodal deep learning"
-abstract: "A flexible package for multimodal-deep-learning to combine tabular data with text and images using Wide and Deep models in Pytorch."
-authors:
-  - name: Javier Rodriguez Zaurin
-    orcid: 0000-0002-1082-1107
-  - name: Pavol Mulinka
-    orcid: 0000-0002-9394-8794
-license: "Apache-2.0 and MIT"
-url: "https://pytorch-widedeep.readthedocs.io/en/latest/index.html"
-repository-code: "https://github.com/jrzaurin/pytorch-widedeep/"
-date-released: 2022-05-18
-keywords:
-  - Python
-  - Pytorch
-  - Deep learning
