### 2023-05-31 #5027 @osorensen
> Now we have this error:
>
> ```
> {
>     "errors": "The license ID you have selected is not present in our system. For the available licenses please check in the following URL https://developers.zenodo.org/#licenses"
> }
> ```
>
> Publishing here is proving to be a challenge. We'll do our best if time permits.

Strange that you get all these errors when creating a Zenodo archive. @openjournals/dev, could you please help?
