### 2023-06-01 #5027 @xuanxu
I'd recommend to update the metadata directly in the Zenodo [page](https://zenodo.org/deposit/7908172), once logged in there is an `Edit` button that allows to update title and author info.
