+++ tests/test_feature_importance/test_transformers_feat_imp.py
@@ -1,8 +1,17 @@
 import numpy as np
 import pandas as pd
+import pytest
 
 from pytorch_widedeep import Trainer
-from pytorch_widedeep.models import WideDeep, TabTransformer
+from pytorch_widedeep.models import (
+    SAINT,
+    WideDeep,
+    FTTransformer,
+    TabFastFormer,
+    TabTransformer,
+    SelfAttentionMLP,
+    ContextAttentionMLP,
+)
 from pytorch_widedeep.preprocessing import TabPreprocessor
 
 np.random.seed(42)
@@ -31,39 +40,114 @@ y_te = target[16:]
 
 # ############################TESTS BEGIN #######################################
 
-with_cls_token = False
-tab_preprocessor = TabPreprocessor(
-    cat_embed_cols=cat_cols,
-    continuous_cols=cont_cols,
-    with_attention=True,
-    with_cls_token=with_cls_token,
+
+def _build_model_for_feat_imp_test(model_name, params):
+    if model_name == "tabtransformer":
+        return TabTransformer(
+            input_dim=6, n_blocks=2, n_heads=2, embed_continuous=True, **params
+        )
+    if model_name == "saint":
+        return SAINT(input_dim=6, n_blocks=2, n_heads=2, **params)
+    if model_name == "fttransformer":
+        return FTTransformer(
+            input_dim=6, n_blocks=2, n_heads=2, kv_compression_factor=1.0, **params
+        )
+    if model_name == "tabfastformer":
+        return TabFastFormer(input_dim=6, n_blocks=2, n_heads=2, **params)
+    if model_name == "self_attn_mlp":
+        return SelfAttentionMLP(input_dim=6, n_blocks=2, n_heads=2, **params)
+    if model_name == "cxt_attn_mlp":
+        return ContextAttentionMLP(input_dim=6, n_blocks=2, **params)
+
+
+@pytest.mark.parametrize("with_cls_token", [True, False])
+@pytest.mark.parametrize(
+    "model_name",
+    [
+        "tabtransformer",
+        "saint",
+        "fttransformer",
+        "tabfastformer",
+        "self_attn_mlp",
+        "cxt_attn_mlp",
+    ],
 )
-X_tr = tab_preprocessor.fit_transform(df_tr).astype(float)
-X_te = tab_preprocessor.transform(df_te).astype(float)
+def test_feature_importances(with_cls_token, model_name):
+    tab_preprocessor = TabPreprocessor(
+        cat_embed_cols=cat_cols,
+        continuous_cols=cont_cols,
+        with_attention=True,
+        with_cls_token=with_cls_token,
+    )
+    X_tr = tab_preprocessor.fit_transform(df_tr).astype(float)
+    X_te = tab_preprocessor.transform(df_te).astype(float)
 
-tab_transformer = TabTransformer(
-    column_idx=tab_preprocessor.column_idx,
-    cat_embed_input=tab_preprocessor.cat_embed_input,
-    continuous_cols=tab_preprocessor.continuous_cols,
-    embed_continuous=True,
-    input_dim=6,
-    n_heads=2,
-    n_blocks=2,
-)
+    params = {
+        "column_idx": tab_preprocessor.column_idx,
+        "cat_embed_input": tab_preprocessor.cat_embed_input,
+        "continuous_cols": tab_preprocessor.continuous_cols,
+    }
 
-model = WideDeep(deeptabular=tab_transformer)
+    tab_model = _build_model_for_feat_imp_test(model_name, params)
 
-trainer = Trainer(
-    model,
-    objective="binary",
-)
+    model = WideDeep(deeptabular=tab_model)
 
-trainer.fit(
-    X_tab=X_tr,
-    target=target,
-    n_epochs=1,
[…]
