### 2023-06-12 ccca550
add default vision models
