### 2023-06-15 #5027 @osorensen
@jrzaurin, what is the exact version of `pytorch-widedeep` at https://zenodo.org/record/7908172? Is it `v1.2.1`, or something else?
