### 2023-06-15 7636035
Unit tested feature importance and made it compatible to the #171 fix. I just want to add a notebook and rready to merge/publish
