### 2023-06-15 #5027 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
